fix issues

+++ internal/k3s/install.go
@@ -26,7 +26,7 @@ var (
 	k3SInstallDir string
 	k3sBinary     string
 
-	k3sBundleRegexp = regexp.MustCompile(`k3s-(.*)\.(tar(\.gz)?)|(tgz)`)
+	k3sBundleRegexp = regexp.MustCompile(`k3s.*\.(tar(\.gz)?)|(tgz)`)
 )
 
 func init() {
@@ -42,6 +42,21 @@ type InstallConfig struct {
 	Hostname    string   // host name of the cluster, optional, default from env
 	NodeIP      string   // node ip to bind on as primary internal ip
 	ExternalIPs []string // list of external ip addresses, optional
+	Debug       bool
+}
+
+func (c InstallConfig) k3sInstallArgs() string {
+	k3sArgs := []string{
+		// binding node to localhost only
+		fmt.Sprintf("--node-ip=%s", c.NodeIP),
+		fmt.Sprintf("--flannel-iface=%s", c.Iface),
+		fmt.Sprintf("--default-local-storage-path=%s", defaultLocalStoragePath),
+	}
+	if len(c.ExternalIPs) > 0 {
+		k3sArgs = append(k3sArgs, fmt.Sprintf("--node-external-ip=%s", strings.Join(c.ExternalIPs, ",")))
+	}
+
+	return strings.Join(k3sArgs, " ")
 }
 
 // Install runs K3S installation process
@@ -54,14 +69,21 @@ func Install(ctx context.Context, c InstallConfig) error {
 		return err
 	}
 
-	name, version, err := findBundle(c.BundlePath)
+	if c.BundlePath != "" {
+		err := bundle.SetBundlePath(c.BundlePath)
+		if err != nil {
+			return err
+		}
+	}
+
+	name, manifest, err := findBundle()
 	if err != nil {
 		return err
 	}
 
-	bundle := bundle.Tar(name)
+	fmt.Printf("Installing K3S %q\n", manifest.K3S)
 
-	fmt.Printf("Installing K3S %q\n", version)
+	bundle := bundle.Tar(name)
 
 	err = errors.Join(
 		bundle.GetFiles(copyK3S, "k3s"),
@@ -70,24 +92,27 @@ func Install(ctx context.Context, c InstallConfig) error {
 	)
 
 	if err != nil {
-		cleanup()
+		cleanup(c.Debug)
 		return err
 	}
 
 	if errors.Is(ctx.Err(), context.Canceled) {
-		cleanup()
+		cleanup(c.Debug)
 		return ctx.Err()
 	}
 
 	return nil
 }
 
-func cleanup() error {
-	exec.Command("k3s-uninstall.sh").Run()
-	os.RemoveAll(k3sImagesPath)
-	os.Remove(k3sBinary)
-	os.Remove(k3sResolvConfPath)
-	return nil
+// cleanup runs k3s-uninstall and removes copied files
+// if debug flag is not enabled
+func cleanup(debug bool) {
+	if !debug {
+		exec.Command("k3s-uninstall.sh").Run()
+		os.RemoveAll(k3sImagesPath)
+		os.Remove(k3sBinary)
+		os.Remove(k3sResolvConfPath)
+	}
 }
 
 func copyK3S(_ fs.FileInfo, r io.Reader) error {
@@ -99,8 +124,7 @@ func copyK3S(_ fs.FileInfo, r io.Reader) error {
 
 	_, err = io.Copy(f, r)
 	if err != nil && !errors.Is(err, io.EOF) {
-		err = errors.Join(err, f.Close())
-		err = errors.Join(err, os.Remove(k3sBinary))
+		err = errors.Join(err, f.Close(), os.Remove(k3sBinary))
 		return err
 	}
 	return nil
@@ -139,18 +163,7 @@ func runInstallScript(ctx context.Context, c InstallConfig) func(fs.FileInfo, io
 		os.Setenv("INSTALL_K3S_SKIP_DOWNLOAD", "true")
 		os.Setenv("INSTALL_K3S_SELINUX_WARN", "true")
 		os.Setenv("INSTALL_K3S_SKIP_SELINUX_RPM", "true")
-
-		k3sArgs := []string{
-			// binding node to localhost only
-			fmt.Sprintf("--node-ip=%s", c.NodeIP),
-			fmt.Sprintf("--flannel-iface=%s", c.Iface),
-			fmt.Sprintf("--default-local-storage-path=%s", defaultLocalStoragePath),
-		}
-		if len(c.ExternalIPs) > 0 {
-			k3sArgs = append(k3sArgs, fmt.Sprintf("--node-external-ip=%s", strings.Join(c.ExternalIPs, ",")))
-		}
-
-		os.Setenv("INSTALL_K3S_EXEC", strings.Join(k3sArgs, " "))
+		os.Setenv("INSTALL_K3S_EXEC", c.k3sInstallArgs())
 
 		cmd := exec.CommandContext(ctx, "sh", "-")
 		cmd.Stdin = r
